Added CoefficientDivision rules for derivatives and summations
Renamed OptimizerRuleSet to OptimizerRuleEngine
Moved OptimizerRuleEngine into its own file
Added appropriate unit tests
